requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
geopy==2.4.0
overpy==0.7
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
httpx==0.27.0
osmnx>=1.9.0
pandas>=2.0.0
shapely>=2.0.0
pyproj>=3.0.0
python-dotenv>=1.0.0